move update integration tests into own class
