+++ sql/src/test/java/io/crate/integrationtests/TransportSQLActionTest.java
@@ -43,7 +43,6 @@ import javax.annotation.Nullable;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
-import static com.google.common.collect.Maps.newHashMap;
 import static org.hamcrest.Matchers.*;
 import static org.hamcrest.core.Is.is;
 
@@ -875,376 +874,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
     }
 
     @Test
-    public void testUpdate() throws Exception {
-        execute("create table test (message string)");
-        ensureGreen();
-
-        execute("insert into test values('hello'),('again')");
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("update test set message='b' where message = 'hello'");
-
-        assertEquals(1, response.rowCount());
-        assertThat(response.duration(), greaterThanOrEqualTo(0L));
-        refresh();
-
-        execute("select message from test where message='b'");
-        assertEquals(1, response.rowCount());
-        assertEquals("b", response.rows()[0][0]);
-
-    }
-
-    @Test
-    public void testUpdateWithExpression() throws Exception {
-        execute("create table test (id integer, other_id long)");
-        ensureGreen();
-
-        execute("insert into test (id, other_id) values(1, 10),(2, 20)");
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("update test set id=(id+10)*other_id");
-
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("select id from test order by id");
-        assertEquals(2, response.rowCount());
-        assertEquals(110, response.rows()[0][0]);
-        assertEquals(240, response.rows()[1][0]);
-    }
-
-    @Test
-    public void testUpdateMultipleDocuments() throws Exception {
-        prepareCreate("test")
-                .addMapping("default",
-                        "message", "type=string,index=not_analyzed")
-                .execute().actionGet();
-        ensureGreen();
-        execute("insert into test values('hello'),('again'),('hello')");
-        assertEquals(3, response.rowCount());
-        refresh();
-
-        execute("update test set message='b' where message = 'hello'");
-
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("select message from test where message='b'");
-        assertEquals(2, response.rowCount());
-        assertEquals("b", response.rows()[0][0]);
-
-    }
-
-    @Test
-    public void testTwoColumnUpdate() throws Exception {
-        prepareCreate("test")
-                .addMapping("default",
-                        "col1", "type=string,index=not_analyzed",
-                        "col2", "type=string,index=not_analyzed")
-                .execute().actionGet();
-        ensureGreen();
-
-        execute("insert into test values('hello', 'hallo'), ('again', 'nochmal')");
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("update test set col1='b' where col1 = 'hello'");
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select col1, col2 from test where col1='b'");
-        assertEquals(1, response.rowCount());
-        assertEquals("b", response.rows()[0][0]);
-        assertEquals("hallo", response.rows()[0][1]);
-
-    }
-
-    @Test
-    public void testUpdateWithArgs() throws Exception {
-        execute("create table test (" +
-                "  coolness float, " +
-                "  details array(object)" +
-                ")");
-        ensureGreen();
-
-        execute("insert into test values(1.1, ?),(2.2, ?)", new Object[]{new Object[0],
-                new Object[]{
-                        new HashMap<String, Object>(),
-                        new HashMap<String, Object>() {{
-                            put("hello", "world");
-                        }}
-                }
-        });
-        assertEquals(2, response.rowCount());
-        refresh();
-
-        execute("update test set coolness=3.3, details=? where coolness = ?", new Object[]{new Object[0], 2.2});
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness from test where coolness=3.3");
-        assertEquals(1, response.rowCount());
-        assertEquals(3.3, response.rows()[0][0]);
-
-    }
-
-    @Test
-    public void testUpdateNestedObjectWithoutDetailedSchema() throws Exception {
-        execute("create table test (coolness object)");
-        ensureGreen();
-
-        Map<String, Object> map = new HashMap<>();
-        map.put("x", "1");
-        map.put("y", 2);
-        Object[] args = new Object[]{map};
-
-        execute("insert into test values (?)", args);
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("update test set coolness['x'] = '3'");
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness['x'], coolness['y'] from test");
-        assertEquals(1, response.rowCount());
-        assertEquals("3", response.rows()[0][0]);
-        assertEquals(2, response.rows()[0][1]);
-    }
-
-    @Test
-    public void testUpdateNestedNestedObject() throws Exception {
-        execute("create table test (" +
-                "coolness object as (x object as (y object as (z int), a string, b string))," +
-                "a object as (x string, y int)," +
-                "firstcol int, othercol int" +
-                ") with (number_of_replicas=0)");
-        ensureGreen();
-
-        Map<String, Object> map = new HashMap<>();
-        map.put("x", "1");
-        map.put("y", 2);
-        Object[] args = new Object[]{map};
-
-        execute("insert into test (a) values (?)", args);
-        refresh();
-
-        execute("update test set coolness['x']['y']['z'] = 3");
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness['x'], a from test");
-        assertEquals(1, response.rowCount());
-        assertEquals("{y={z=3}}", response.rows()[0][0].toString());
-        assertEquals(map, response.rows()[0][1]);
-
-        execute("update test set firstcol = 1, coolness['x']['a'] = 'a', coolness['x']['b'] = 'b', othercol = 2");
-        assertEquals(1, response.rowCount());
-        refresh();
-        waitNoPendingTasksOnAll();
-
-        execute("select coolness['x']['b'], coolness['x']['a'], coolness['x']['y']['z'], " +
-                "firstcol, othercol from test");
-        assertEquals(1, response.rowCount());
-        Object[] firstRow = response.rows()[0];
-        assertEquals("b", firstRow[0]);
-        assertEquals("a", firstRow[1]);
-        assertEquals(3, firstRow[2]);
-        assertEquals(1, firstRow[3]);
-        assertEquals(2, firstRow[4]);
-    }
-
-    @Test
-    public void testUpdateNestedObjectDeleteWithArgs() throws Exception {
-        execute("create table test (a object as (x object as (y int, z int))) with (number_of_replicas=0)");
-        ensureGreen();
-
-        Map<String, Object> map = newHashMap();
-        Map<String, Object> nestedMap = newHashMap();
-        nestedMap.put("y", 2);
-        nestedMap.put("z", 3);
-        map.put("x", nestedMap);
-        Object[] args = new Object[]{map};
-
-        execute("insert into test (a) values (?)", args);
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("update test set a['x']['z'] = ?", new Object[]{null});
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select a['x']['y'], a['x']['z'] from test");
-        assertEquals(1, response.rowCount());
-        assertEquals(2, response.rows()[0][0]);
-        assertNull(response.rows()[0][1]);
-    }
-
-    @Test
-    public void testUpdateNestedObjectDeleteWithoutArgs() throws Exception {
-        execute("create table test (a object as (x object as (y int, z int))) with (number_of_replicas=0)");
-        ensureGreen();
-
-        Map<String, Object> map = newHashMap();
-        Map<String, Object> nestedMap = newHashMap();
-        nestedMap.put("y", 2);
-        nestedMap.put("z", 3);
-        map.put("x", nestedMap);
-        Object[] args = new Object[]{map};
-
-        execute("insert into test (a) values (?)", args);
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("update test set a['x']['z'] = null");
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select a['x']['z'], a['x']['y'] from test");
-        assertEquals(1, response.rowCount());
-        assertNull(response.rows()[0][0]);
-        assertEquals(2, response.rows()[0][1]);
-    }
-
-    @Test
-    public void testUpdateWithNestedObjectArrayIdxAccess() throws Exception {
-        execute("create table test (coolness array(float)) with (number_of_replicas=0)");
-        ensureGreen();
-        execute("insert into test values (?)", new Object[]{new Object[]{2.2, 2.3, 2.4}});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        expectedException.expect(SQLActionException.class);
-        expectedException.expectMessage("Array index must be in range 1 to 2147483648");
-
-        execute("update test set coolness[0] = 3.3");
-    }
-
-    @Test
-    public void testUpdateNestedObjectWithDetailedSchema() throws Exception {
-        execute("create table test (coolness object as (x string, y string))");
-        ensureGreen();
-        Map<String, Object> map = new HashMap<>();
-        map.put("x", "1");
-        map.put("y", "2");
-        Object[] args = new Object[]{map};
-
-        execute("insert into test values (?)", args);
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("update test set coolness['x'] = '3'");
-
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness from test");
-        assertEquals(1, response.rowCount());
-        assertEquals("{y=2, x=3}", response.rows()[0][0].toString());
-    }
-
-    @Test
-    public void testUpdateResetNestedObject() throws Exception {
-        prepareCreate("test")
-                .addMapping("default",
-                        "coolness", "type=object,index=not_analyzed")
-                .execute().actionGet();
-        ensureGreen();
-
-        Map<String, Object> map = new HashMap<>();
-        map.put("x", "1");
-        map.put("y", 2);
-        Object[] args = new Object[]{map};
-
-        execute("insert into test values (?)", args);
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        // update with different map
-        Map<String, Object> new_map = new HashMap<>();
-        new_map.put("z", 1);
-
-        execute("update test set coolness = ?", new Object[]{new_map});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness from test");
-        assertEquals(1, response.rowCount());
-        assertEquals(new_map, response.rows()[0][0]);
-
-        // update with empty map
-        Map<String, Object> empty_map = new HashMap<>();
-
-        execute("update test set coolness = ?", new Object[]{empty_map});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness from test");
-        assertEquals(1, response.rowCount());
-        assertEquals(empty_map, response.rows()[0][0]);
-    }
-
-    @Test
-    public void testUpdateResetNestedObjectUsingUpdateRequest() throws Exception {
-        XContentBuilder mapping = XContentFactory.jsonBuilder().startObject()
-                .startObject("default")
-                .startObject("_meta").field("primary_keys", "id").endObject()
-                .startObject("properties")
-                .startObject("id")
-                .field("type", "string")
-                .field("index", "not_analyzed")
-                .endObject()
-                .startObject("data")
-                .field("type", "object")
-                .field("index", "not_analyzed")
-                .field("dynamic", false)
-                .endObject()
-                .endObject()
-                .endObject()
-                .endObject();
-
-        prepareCreate("test")
-                .addMapping("default", mapping)
-                .execute().actionGet();
-        ensureGreen();
-
-        Map<String, Object> data = new HashMap<String, Object>() {{
-            put("foo", "bar");
-            put("days", new ArrayList<String>() {{
-                add("Mon");
-                add("Tue");
-                add("Wen");
-            }});
-        }};
-        execute("insert into test (id, data) values (?, ?)", new Object[]{"1", data});
-        refresh();
-
-        execute("select data from test where id = ?", new Object[]{"1"});
-        assertEquals(data, response.rows()[0][0]);
-
-        Map<String, Object> new_data = new HashMap<String, Object>() {{
-            put("days", new ArrayList<String>() {{
-                add("Mon");
-                add("Wen");
-            }});
-        }};
-        execute("update test set data = ? where id = ?", new Object[]{new_data, "1"});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select data from test where id = ?", new Object[]{"1"});
-        assertEquals(new_data, response.rows()[0][0]);
-    }
-
-    @Test
     public void testGetResponseWithObjectColumn() throws Exception {
         XContentBuilder mapping = XContentFactory.jsonBuilder().startObject()
                 .startObject("default")
@@ -1275,38 +904,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
 
         execute("select data from test where id = ?", new Object[]{"1"});
         assertEquals(data, response.rows()[0][0]);
-    }
-
-    @Test
-    public void testUpdateResetNestedNestedObject() throws Exception {
-        prepareCreate("test")
-                .addMapping("default",
-                        "coolness", "type=object,index=not_analyzed")
-                .execute().actionGet();
-        ensureGreen();
-
-        Map<String, Object> map = new HashMap<String, Object>() {{
-            put("x", "1");
-            put("y", new HashMap<String, Object>() {{
-                put("z", 3);
-            }});
-        }};
-
-        execute("insert into test values (?)", new Object[]{map});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        Map<String, Object> new_map = new HashMap<>();
-        new_map.put("a", 1);
-
-        execute("update test set coolness['y'] = ?", new Object[]{new_map});
-        assertEquals(1, response.rowCount());
-        refresh();
-
-        execute("select coolness['y'], coolness['x'] from test");
-        assertEquals(1, response.rowCount());
-        assertEquals(new_map, response.rows()[0][0]);
-        assertEquals("1", response.rows()[0][1]);
     }
 
     @Test
@@ -1402,25 +999,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
 
         execute("select * from test where pk_col='123'");
         assertEquals(0, response.rowCount());
-    }
-
-    @Test
-    public void testUpdateToUpdateRequestByPlanner() throws Exception {
-        this.setup.createTestTableWithPrimaryKey();
-
-        execute("insert into test (pk_col, message) values ('123', 'bar')");
-        assertEquals(1, response.rowCount());
-        waitNoPendingTasksOnAll(); // wait for new columns to be available
-        refresh();
-
-        execute("update test set message='bar1' where pk_col='123'");
-        assertEquals(1, response.rowCount());
-        assertThat(response.duration(), greaterThanOrEqualTo(0L));
-        refresh();
-
-        execute("select message from test where pk_col='123'");
-        assertEquals(1, response.rowCount());
-        assertEquals("bar1", response.rows()[0][0]);
     }
 
     @Test
@@ -1904,45 +1482,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
     }
 
     @Test
-    public void testUpdateByIdWithMultiplePrimaryKeyAndClusteredBy() throws Exception {
-        execute("create table quotes (id integer primary key, author string primary key, " +
-                "quote string) clustered by(author) with (number_of_replicas=0)");
-        ensureGreen();
-        execute("insert into quotes (id, author, quote) values(?, ?, ?)",
-                new Object[]{1, "Ford", "I'd far rather be happy than right any day."});
-        assertEquals(1L, response.rowCount());
-
-        execute("update quotes set quote=? where id=1 and author='Ford'",
-                new Object[]{"Don't panic"});
-        assertEquals(1L, response.rowCount());
-        refresh();
-
-        execute("select quote from quotes where id=1 and author='Ford'");
-        assertEquals(1L, response.rowCount());
-        assertThat((String) response.rows()[0][0], is("Don't panic"));
-    }
-
-    @Test
-    public void testUpdateByQueryWithMultiplePrimaryKeyAndClusteredBy() throws Exception {
-        execute("create table quotes (id integer primary key, author string primary key, " +
-                "quote string) clustered by(author) with (number_of_replicas=0)");
-        ensureGreen();
-        execute("insert into quotes (id, author, quote) values(?, ?, ?)",
-                new Object[]{1, "Ford", "I'd far rather be happy than right any day."});
-        assertEquals(1L, response.rowCount());
-        refresh();
-
-        execute("update quotes set quote=? where id=1",
-                new Object[]{"Don't panic"});
-        assertEquals(1L, response.rowCount());
-        refresh();
-
-        execute("select quote from quotes where id=1 and author='Ford'");
-        assertEquals(1L, response.rowCount());
-        assertThat((String) response.rows()[0][0], is("Don't panic"));
-    }
-
-    @Test
     public void testDeleteByIdWithMultiplePrimaryKey() throws Exception {
         execute("create table quotes (id integer primary key, author string primary key, " +
                 "quote string) with (number_of_replicas=0)");
@@ -2105,43 +1644,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
         assertThat((String) response.rows()[2][0], is("Algol is a Star System"));
         // ...
 
-    }
-
-    @Test
-    public void testUpdateVersionHandling() throws Exception {
-        execute("create table test (id int primary key, c int) with (number_of_replicas=0, refresh_interval=0)");
-        ensureGreen();
-        execute("insert into test (id, c) values (1, 1)");
-        execute("refresh table test");
-        execute("select _version, c from test");
-
-        long version = (Long) response.rows()[0][0];
-        assertThat(version, is(1L));
-
-        // with primary key optimization:
-
-        execute("update test set c = 2 where id = 1 and _version = 1"); // this one works
-        assertThat(response.rowCount(), is(1L));
-        execute("update test set c = 3 where id = 1 and _version = 1"); // this doesn't
-        assertThat(response.rowCount(), is(0L));
-
-        execute("refresh table test");
-        execute("select _version, c from test");
-        assertThat((Long) response.rows()[0][0], is(2L));
-        assertThat((Integer) response.rows()[0][1], is(2));
-
-
-        // without primary key optimization:
-        execute("update test set c = 4 where _version = 2"); // this one works
-        assertThat(response.rowCount(), is(1L));
-        execute("update test set c = 5 where _version = 2"); // this doesn't
-        assertThat(response.rowCount(), is(0L));
-
-
-        execute("refresh table test");
-        execute("select _version, c from test");
-        assertThat((Long) response.rows()[0][0], is(3L));
-        assertThat((Integer) response.rows()[0][1], is(4));
     }
 
     @Test
